feat: #3 Add SettingRoute Message Bodys

+++ include/JT808/MessageBody/DeletingAreaRoute.h
@@ -9,7 +9,7 @@ class DeletingAreaRoute : public MessageBodyBase
 {
 public:
     DeletingAreaRoute() = default;
-    DeletingAreaRoute(const std::vector<uint16_t>& ids);
+    DeletingAreaRoute(const std::vector<uint32_t>& ids);
     void parse(const std::vector<uint8_t>& data) override;
     void parse(const uint8_t* data, int size) override;
     std::vector<uint8_t> package() override;
@@ -17,12 +17,12 @@ public:
 
     uint8_t length() const;
     void setLength(uint8_t newLength);
-    std::vector<uint16_t> ids() const;
-    void setIds(const std::vector<uint16_t>& newIds);
+    std::vector<uint32_t> ids() const;
+    void setIds(const std::vector<uint32_t>& newIds);
 
 private:
     uint8_t m_length = 0;
-    std::vector<uint16_t> m_ids;
+    std::vector<uint32_t> m_ids;
 };
 
 }
